deleted questions components
