+++ code/src/App.js
@@ -1,38 +1,87 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { FormHeader } from './Components/FormHeader'
-import { QuestionOne } from './Components/QuestionOne'
-import { QuestionTwo } from './Components/QuestionTwo'
-import { QuestionThree } from './Components/QuestionThree'
-import { QuestionFour } from './Components/QuestionFour'
-import { SubmitButton } from './Components/SubmitButton'
 import { Summary } from './Components/Summary'
 
 
 
-
+const socialMedias = ["Facebook", "Instagram", "Linkedin", "Others"];
+const programingSkills = ["HTML/CSS", "JavaScript", "Paython"];
 
 
 export const App = () => {
+  const [name, setName] = useState("");
+  const [socialMedia, setSocialMedia] = useState("");
+  const [skills, setSkills] = useState();
+  const [wantsNewsletter, setWantdNewsletter] = useState(false);
+  const [submitted, setSubmitted] = useState(false);
+
   return (
-    <div>
-      <section className="form-container">
+    <div className="form-container">
 
-        <div className="form-wrapper">
+      <div className="form-wrapper">
+
+        <form onSubmit={event => event.preventDefault()}>
+
+          <FormHeader />
           <div>
-            <FormHeader />
-            <QuestionOne />
-            <QuestionTwo />
-            <QuestionThree />
-            <QuestionFour />
-            <SubmitButton />
+            <p>What's your name?</p>
+
+            <input className="name-input"
+              type="text"
+              onChange={event => setName(event.target.value)}
+              value={name}
+            />
           </div>
           <div>
-            <Summary />
+            <p>How did you find about Technigo Bootcamp?</p>
+
+            <select className="drop-down" onChange={event => setSocialMedia(event.target.value)}
+              checked={socialMedia === socialMedias}>
+
+              <option value="">Select one</option>
+
+              {socialMedias.map(item => (
+                <option key={item} value={item}>{item}</option>
+              )
+              )
+              }
+            </select>
+
           </div>
-        </div>
 
-      </section>
+          <div>
 
+            <p>Experience in programming?</p>
+
+            {programingSkills.map(skill => (
+
+              <label key={skill}>
+
+                <input className="radio-circle"
+                  type="radio"
+                  value={skill}
+                  onChange={event => setSkills(event.target.value)}
+                  checked={skills === skill}
+                />
+                {skill}
+              </label>
+            )
+            )
+            }
+          </div>
+
+          <label>
+            <p> I want to receive newsletter</p>
+            <input
+              type="checkbox"
+              checked={wantsNewsletter}
+              onChange={event => setWantdNewsletter(event.target.checked)}
+            />
+          </label>
+          <button onClick={() => setSubmitted(true)}>SUBMIT</button>
+        </form>
+        {submitted && <Summary name={name} socialMedia={socialMedia} skills={skills} wantsNewsletter={wantsNewsletter} />}
+      </div>
     </div>
-  )
-}
+  );
+};
